+++ src/colinux/user/pidfile.c
@@ -34,7 +34,7 @@ co_rc_t read_pid_from_file(co_pathname_t pathname, co_id_t *id)
 
 	if (size > 0) {
 		*id = atoi(buf);
-		co_debug("pid %d from file", *id);
+		co_debug("pid %d from file", (int)*id);
 	} else {
 		co_debug("file '%s' to small or unreadable\n", pathname);
 		rc = CO_RC(INVALID_PARAMETER);
